feat: add metadata to comment
